backend/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Cristian-Jara", email = "cristian.jaras@sansano.usm.cl"},
]
dependencies = [
    "django>=4",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.1",
    "django-filter>=24.1",
    "drf-writable-nested>=0.7.0",
    "django-cors-headers>=4.3.1",
    "python-decouple>=3.8",
    "djoser>=2.2.2",
    "django-cleanup>=8.1.0",
    "psycopg2>=2.9.9",
    "gdal>=3.8.4",
    "djangorestframework-gis>=1.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
start = "python manage.py runserver"
post_init = { composite = ["pdm install", "migrate", "start"] }
test = "python manage.py test"

[tool.pdm.dev-dependencies]
dev = [
    "setuptools>=69.1.1",
    "prospector>=1.10.3",
    "bandit>=1.7.8",
    "pyroma>=4.2",
]
